Add user authentication
Email and password for users.
JSON Web Tokens (JWT).
Logging users in.
React auth context.
Login and signup forms.
Making authorized requests.
Protecting API routes.

+++ wine-one-frontend/src/components/UserDetails.js
@@ -1,4 +1,5 @@
 import { useUsersContext } from '../hooks/useUsersContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
@@ -6,10 +7,18 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 const UserDetails = ({ user }) => {
 
   const {dispatch} = useUsersContext();
+  const {user: _user} = useAuthContext();
   
   const handleClick = async () => {
+    if (!_user) {
+      return;
+    }
+
     const response = await fetch('/api/users/' + user._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${_user.token}`
+      }
     })
     
     const json = await response.json();
